requirements.txt
fastapi
uvicorn
deepface
opencv-python
numpy
gspread
google-auth
python-multipart
faiss-cpu
mediapipe
